# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'QASA API Documentation'
  description: 'API documentation for QASA Laravel application - invoicing and client management system'
  contact:
    name: 'QASA Support'
    email: support@qasa.sk
  license:
    name: MIT
    url: 'https://opensource.org/licenses/MIT'
  version: 1.0.0
paths:
...
  '/api/v1/orders/{order}/items':
    get:
      tags:
        - 'Order Items'
      summary: 'List order items'
      operationId: d1479b5b4674b82a11e62f14b5f9f33d
      parameters:
        -
          name: order
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: 'List of order items'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderItem'
        401:
          description: Unauthenticated
        403:
          description: Forbidden
      security:
        -
          sanctum: []
    post:
      tags:
        - 'Order Items'
      summary: 'Create order item'
      operationId: 175c82292fbf08bb745b1804b7dfc6fd
      parameters:
        -
          name: order
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - type
                - description
                - quantity
                - unit
                - unit_price
                - vat_rate
              properties:
                type:
                  type: string
                  enum:
                    - service
                    - product
                description:
                  type: string
                  example: 'Consulting services'
                  maxLength: 500
                quantity:
                  type: number
                  format: float
                  example: 8.5
                unit:
                  type: string
                  example: h
                  maxLength: 20
                unit_price:
                  type: number
                  format: float
                  example: 50
                vat_rate:
                  description: 'VAT rate in percent (0-100)'
                  type: number
                  format: float
                  example: 20
                sort_order:
                  type: integer
                  example: 1
                  nullable: true
              type: object
      responses:
        201:
          description: 'Order item created'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderItem'
        401:
          description: Unauthenticated
        403:
          description: Forbidden
        422:
          description: 'Validation error'
      security:
        -
          sanctum: []
  '/api/v1/orders/{order}/items/{item}':
    put:
      tags:
        - 'Order Items'
      summary: 'Update order item'
      operationId: fbdb9c9bfe9e78286d5f9729941ea834
      parameters:
        -
          name: order
          in: path
          required: true
          schema:
            type: string
            format: uuid
        -
          name: item
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - type
                - description
                - quantity
                - unit
                - unit_price
                - vat_rate
              properties:
                type:
                  type: string
                  enum:
                    - service
                    - product
                description:
                  type: string
                  example: 'Consulting services'
                  maxLength: 500
                quantity:
                  type: number
                  format: float
                  example: 8.5
                unit:
                  type: string
                  example: h
                  maxLength: 20
                unit_price:
                  type: number
                  format: float
                  example: 50
                vat_rate:
                  description: 'VAT rate in percent (0-100)'
                  type: number
                  format: float
                  example: 20
                sort_order:
                  type: integer
                  example: 1
                  nullable: true
              type: object
      responses:
        200:
          description: 'Order item updated'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderItem'
        401:
          description: Unauthenticated
        403:
          description: Forbidden
        422:
          description: 'Validation error'
      security:
        -
          sanctum: []
    delete:
      tags:
        - 'Order Items'
      summary: 'Delete order item'
      operationId: 0483db89e6303953af461e8407737db3
      parameters:
        -
          name: order
          in: path
          required: true
          schema:
            type: string
            format: uuid
        -
          name: item
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        204:
          description: 'Order item deleted'
        401:
          description: Unauthenticated
        403:
          description: Forbidden
        422:
          description: 'Cannot delete invoiced item'
      security:
        -
          sanctum: []
  '/api/v1/orders/{order}/notes':
    get:
      tags:
        - 'Order Notes'
      summary: 'List order notes'
      operationId: bde7e4b4d10027ced86a278eb3462bc9
      parameters:
        -
          name: order
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: 'List of order notes'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderNote'
        401:
          description: Unauthenticated
        403:
          description: Forbidden
      security:
        -
          sanctum: []
    post:
      tags:
        - 'Order Notes'
      summary: 'Create order note'
      operationId: d6e631d850a9ab47be4a84a89b3d1d06
      parameters:
        -
          name: order
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - content
              properties:
                content:
                  type: string
                  example: 'Customer requested expedited delivery'
              type: object
      responses:
        201:
          description: 'Order note created'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderNote'
        401:
          description: Unauthenticated
        403:
          description: Forbidden
        422:
          description: 'Validation error'
      security:
        -
          sanctum: []
  '/api/v1/orders/{order}/notes/{note}':
    delete:
      tags:
        - 'Order Notes'
      summary: 'Delete order note'
      operationId: 207caaf0f89130f40a3b385d24d3cf6a
      parameters:
        -
          name: order
          in: path
          required: true
          schema:
            type: string
            format: uuid
        -
          name: note
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        204:
          description: 'Order note deleted'
        401:
          description: Unauthenticated
        403:
          description: 'Forbidden or only author and managers can delete'
      security:
        -
          sanctum: []
...
components:
  schemas:
...
    OrderItem:
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum:
            - service
            - product
        description:
          type: string
        quantity:
          type: number
          format: float
        unit:
          type: string
          example: ks
        unit_price:
          type: number
          format: float
        vat_rate:
          type: number
          format: float
        vat_amount:
          type: number
          format: float
        total_excl_vat:
          type: number
          format: float
        total_incl_vat:
          type: number
          format: float
        sort_order:
          type: integer
        created_at:
          type: string
          format: date-time
          nullable: true
      type: object
    OrderNote:
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
        created_at:
          type: string
          format: date-time
          nullable: true
        updated_at:
          type: string
          format: date-time
          nullable: true
      type: object
...
  securitySchemes:
    sanctum:
      type: apiKey
      description: 'Enter token in format (Bearer <token>)'
      name: Authorization
      in: header
...
security:
  -
    sanctum: []